nucleus: move "real" core runtime to runtime.c and process.c
Separate ecl_core from cl_core as representing the common language runtime
state (as opposed to common lisp state).

The first environment is statically allocated for simplicity. Managing
process-wide environments is factored out from the threads module.

This commit also introduces a preliminary support for switching global
runtime environment in single-threaded builds.

ecl_thread_internal_error is now an alias to ecl_internal_error and causes the
process to abort. IMHO it is justified to punt on internal thread errors. This
is to ensure, that process.c does not depend on pthread_exit that is replaced by
BoehmGC with GC_pthread_exit forcing us to link nucleus against libgc.