subtypep: introduce the function MAKE-REGISTERED-TAG
This function is used by

- register-elementary-interval
- register-type

Additionally we drop logandc2 in the invocation to update type, because
FIND-TYPE-BOUNDS always does that for us (it was redundant).

Additionally imrove some comments and replace LET* with LET where suitable.
Numeric range predicate and comparator are defined as their own functions
instead of lambda expressions.