subtypep: don't add a new tag for equivalent types
The function MAKE-REGISTERED-TAG calls FIND-TYPE-BOUNDS to determine supertypes
that need to be updated with the new bit, and subtypes that need to be included
in the new tag.

Thiis procedure was bogus because it did not recognize equivalent types. That
lead to a situation, where synonymous types could have been added twice with
incorrect relation. Consider:

type A: 011
type B: 001

We add a type C that is equivalent to A and B is subtype (to both). With the old
method the result would be:

type A: 111
type B: 001
type C: 101

So if we had later queried wheter A is subtypep to C, then the answer would be
incorrectly NIL.

The bug was hidden by the fact, that CANONICAL-TYPE expands type aliases when
they are symbols, so we had never encountered a situation where equivalent types
had different names in *ELEMENTARY-TYPES*. This changes when we introduce the
new kingdom for CONS type, because the key is (CONS X Y), and symbols in type
names X Y are not expanded, so

  (CONS (OR FIXNUM BIGNUM)) is not EQUAL to (CONS INTEGER)