speed up looking up and setting values of dynamic variables
Introduce ecl_symbol_value_nonnull and ecl_setq_nonnull which do the
minimal amount of work needed to implement symbol-value and setq for
dynamic variables which we have checked to be non-null and of type
symbol in the compiler.

Also introduce a type check in ecl_symbol_value to be consistent with
ecl_setq which also checks the type. These two functions are mainly
used for the embedding interface now, so for that reason it is also
useful to have a type check in there.