cmp: cleanup: remove a misleading comment for VAR-LOC
In the past VAR-LOC in the first pass could have a value 'CLB to indicate, that
it can't be allocated on the C stack. But somewhere along the way the structure
gained a separate slot REF-CLB that indicates that and VAR-LOC in the 1st pass
is /always/ OBJECT.

Update comments to reflect that fact and remove no-op assignments.