cmp: exit manager: introduce operators UNWIND-FLEE and UNWIND-COND
The operator UNWIND-FLEE is used to perform a dynamic unwind. Previously we've
opencoded this type of exits in appropriate operators.

The operator UNWIND-COND is used to perform a conditional unwind. It is expected
to be called to produce the IF statement body. With this commit all every
transfer of control goes through the exit manager.

Ultimately we will want to include the if test directly in UNWIND-COND.