cmp: variable holding function has gensym as a name
Previously we've named the function's variable with the function name. That
could lead to a varible named (SETF FOO), and that can break down the road
because variable names are assumed to be symbols.

Moreover this change will prevent a possible fail scenario where we inline a
constant value with a known name instead of the function with the same name.