cmp: improvements for (c1nil) and (c1t)
Make the types more precise to improve type propagation. We will fall
back to using object-type as appropriate in later stages of the
compilation.

Also return a new copy for each invocation instead of the same *c1nil*
and *c1t* since we may later set the type of the c1form (e.g. in
enforce-types). This can happen for invalid code which can permanently
set the type of *c1nil* to NIL, leading to miscompilations later on
for unrelated valid code.