process: don't call cl_sleep on a variable declared static
We call ecl_musleep that avoids the hassle of converting values back and forth
between lisp and c world.

The change is prompted by a potential bug in cosmopolitan libc, where the value
stored in a static variable can't be passed to a function in a different
compilation unit. Unmodified code lead to a segmentation fault with cosmocc.