ecl_interpreter: abstract away all lcl accessors
We are going to change the representation of the local environment, but first we
make identify accessors and put them behind macros.

While doing so the OP_LABELS has been changed to look similar to OP_FLET. Among
other things we cons separately functions into fun_env, but this inefficiency
will be removed later when we address local entries from the frame.base.