internal: ignore array-bounds errors on GCC >= 12
As noted in the comment, GCC 12.0 onwards generates an invalid warning about
array bounds when we cast a stack allocated object to cl_object and use it

    warning: array subscript ‘union cl_lispunion[0]’ is partly outside array
    bounds of ‘struct ecl_stack_frame[1]’ [-Warray-bounds=]

The code is conforming, but apparently GCC IR has generated a dead branch that
accesses the array out of bounds. For time being we disable this warning.

See also:
https://gitlab.com/libeigen/eigen/-/issues/2506
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=106274
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=105523

and generally explroe gcc bugzilla for this kind of false positive.