stack frames: don't displace onto env->values vector
ad6f1f7f10e66069207dac01ff6f83406e0aced4 (2008-05-12) allowed for displacing
stack frames onto the values vector. This commit seems to be before we've stored
them on lisp stack.

As for the rationale why we revert that change:

1. There may be a slight performance improvement because the standard gf
dispatch won't be forced to copy the stack frame

2. Stack frame operators assume that they are displaced onto env->stack*, for
example ecl_stack_frame_push increments env->stack_top

3. If we fully buy into assumption that the stack frame is displaced onto
env->stack* then we may replace pointers with indexes and protect the user from
accessing invalidated memory when the stack grows

The last point will be implemented in the next commit. It is worth noting, that
frames auto-heal themselves if their own operation is the cause of the stack
growth, but all other stack frames are left invalid after the operation.