[wrong] boot: move ecl_core_struct to boot.d
This commit "moves" ecl_core struct to boot.d, but also modifies semantics by
statically allocating the first environment (see #828). So this commit needs to
be reworked to be faithful in moving it, and eventually after that change
semantics (but, again, see #828).

IIRC moving the code broke building nucleus, but I don't remember details; this
issue needs to be addressed (without breaking #828) before this commit is added
for inclusion.