core: fix a race condition between core symbols and the gc
The GC is enabled when we create initial symbols. The function make_this_symbol
initializes symbols from a static table as follows:

1. initialize symbol structure fields
2. associate the symbol with its package
3. associate the symbol with its function
4. increment cl_num_symbols_in_core

Most notably in point 2:
s->symbol.name = ecl_make_constant_base_string(name,-1);

GC's stacks_scanner marks

  [cl_symbols; cl_symbols+cl_num_symbols_in_core]

in other words, if the GC kicks in between 1. and 4., the symbol name may be
garbage collected, messing the symbol name and package exports.

To mitigate that, we assign intermediate values on the stack and increment
the number of symbols in core after the symbol is initialized:

1. store initialization values on stack -- cl_object foo = alloc();
2. initialize symbol structure fields
3. increment cl_num_symbols_in_core
4. associate the symbol with its package and function