bytecmp: lazily allocate a vector for captured env objects
This way we don't need to separately keep track for whether the function
boundary was crossed for the env, and we don't need to /think hard/ when
captured vector should be allocated.

Moreover the test for whether the parent crosses the function boundary is
precise now, and we save memory when we don't need to allocat a vector.