predlib: add accessors to elements of *ELEMENTARY-TYPES*
Previously elementary types were considered to be (CONS NAME TAG), but I want to
introduce additional slot information to them, so we define a structure for that
type. The representation a list because MAYBE-SAVE-TYPES calls COPY-TREE. Also
DEFSTRUCT is not available yet.

Rename PUSH-TYPE to PUSH-NEW-TYPE and move it to a correct section in the file.