cmp: fix constant folding
Fixes a regression introduced in commit
816c08340bbd161fed82c0d625073f3195d685c3. After the changes in that
commit, c1forms of name variable can denote constants as well.

Since we are now more aggressive with inlining constant access, we
have to fix a bad defconstant definition in clos/boot.lsp as well (the
new definition makes sure that the compile-time and load-time
evaluations of the constant are eql).